Run a broader set of tests on travis
Includes jws online tests, and different rubies, parallelized between suites. Coverall is switched off.
This forms an experimental branch for running a deeper set of tests that will be merged into periodically but not used directly. It allows the main master branch to collect stats for Coverall and code-climate, and run only with the target 2.1.2 ruby.